services/ingest/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0
redis==7.0.1
python-dotenv==1.0.1
